⚡ Moved routes to routes.ts file ⚡ Fixed AuthGuard Bug ⚡ Created AuthService ⚡ Added Confirm in Movie Delete ⚡ Added footer with project link

+++ src/app/containers/edit-movie.component.ts
@@ -61,7 +61,9 @@ export class EditMovieComponent implements OnInit {
   }
 
   delete() {
-    this.appService.deleteMovie(this.movieId).pipe(take(1)).subscribe();
-    this.router.navigate([ '/movies' ]);
+    if (confirm('Are you sure you want to delete this movie?')) {
+      this.appService.deleteMovie(this.movieId).pipe(take(1)).subscribe();
+      this.router.navigate([ '/movies' ]);
+    }
   }
 }
